+++ src/Pong/update.js
@@ -1,4 +1,4 @@
-import { WINDOW_HEIGHT } from "../Pong/constants";
+import { WINDOW_HEIGHT, WINDOW_WIDTH, BALL_SPEED } from "../Pong/constants";
 
 export const paddleAuto = (ball, paddle) => {
     if(ball.velocity.x > 0) {
@@ -20,7 +20,61 @@ export const paddleAuto = (ball, paddle) => {
     }
 }
 
-export const updateBall = (ball) => {
+export const updateBall = ball => {
     ball.position.x += ball.velocity.x;
     ball.position.y += ball.velocity.y;
+}
+
+export const update = (refBall, refGame, refPaddlePlayer2, refWinCondition) => {
+    const ball = refBall.current;
+    const score = refGame.current.score;
+    const speed = refGame.current.ballSpeed;
+    const paddlePlayer2 = refPaddlePlayer2.current;
+    const winCondition = refWinCondition.current;
+
+    
+    updateBall(refBall.current)
+    
+    paddleAuto(ball, paddlePlayer2);
+
+    //Update score and reset ball position and velocity
+    const  initialPosition = WINDOW_HEIGHT / 2 - 50 + Math.random() * 100; 
+    const angle = Math.random() * Math.PI / 4;
+    const sign = Math.random()< 0.5 ? -1 : 1;
+  
+    if(ball.position.x <= 0){
+        score.player2++;
+        refGame.current.ballSpeed = BALL_SPEED;
+        ball.position.x = WINDOW_WIDTH / 2;
+        ball.position.y = initialPosition;
+        ball.velocity.x = -speed * Math.cos(angle);
+        ball.velocity.y = sign * speed * Math.sin(angle);
+        
+    }
+
+    if(ball.position.x >= WINDOW_WIDTH){
+        score.player1++;
+        refGame.current.ballSpeed = BALL_SPEED;
+        ball.position.x = WINDOW_WIDTH / 2;
+        ball.position.y = WINDOW_HEIGHT / 2;
+        ball.velocity.x = speed * Math.cos(angle);
+        ball.velocity.y = sign * speed * Math.sin(angle);
+    }
+
+    //Check if winning condition achieved;
+    if(score.player1 > 3 && score.player1 > score.player2 + 1){
+        winCondition.gameOver = true;
+        refGame.current.screen = "gameOver";
+        winCondition.winner = "Player 1";
+        ball.velocity.x = 0;
+        ball.velocity.y = 0;
+    }
+
+    if(score.player2 > 3 && score.player2 > score.player1 + 1){
+        winCondition.gameOver = true;
+        refGame.current.screen = "gameOver";
+        winCondition.winner = "Player 2";
+        ball.velocity.x = 0;
+        ball.velocity.y = 0;
+    }
 }
